wip - generic data model
