+++ src/main/java/de/zalando/zmon/dataservice/data/KairosDbWorkResultWriter.java
@@ -1,5 +1,6 @@
 package de.zalando.zmon.dataservice.data;
 
+import de.zalando.zmon.dataservice.config.DataServiceConfigProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,8 +11,10 @@ import com.codahale.metrics.Timer;
 
 import de.zalando.zmon.dataservice.DataServiceMetrics;
 
+import java.util.List;
+
 @Component
-public class KairosDbWorkResultWriter implements WorkResultWriter {
+public class KairosDbWorkResultWriter extends AbstractWorkResultWriter {
 
     public static final String KAIROS_WRITER_EXECUTOR = "kairos-writer";
     private final Logger log = LoggerFactory.getLogger(KairosDbWorkResultWriter.class);
@@ -21,27 +24,27 @@ public class KairosDbWorkResultWriter implements WorkResultWriter {
     private final KairosDBStore kairosStore;
 
     @Autowired
-    KairosDbWorkResultWriter(KairosDBStore kairosStore, DataServiceMetrics metrics) {
+    KairosDbWorkResultWriter(DataServiceConfigProperties config,
+                             KairosDBStore kairosStore,
+                             DataServiceMetrics metrics) {
+        super(config);
         this.kairosStore = kairosStore;
         this.metrics = metrics;
     }
 
     @Async(KAIROS_WRITER_EXECUTOR)
     @Override
-    public void write(WriteData writeData) {
-        log.debug("write to KairosDB ...");
-        if (writeData.getWorkerResultOptional().isPresent()) {
-            Timer.Context c = metrics.getKairosDBTimer().time();
-            try {
-                kairosStore.store(writeData.getWorkerResultOptional().get());
-                log.debug("... written to KairosDb");
-            } catch (Exception e) {
-                log.error("failed kairosdb write check={} data={}", writeData.getCheckId(), writeData.getData(), e);
-                metrics.markKairosError();
-            } finally {
-                c.stop();
-            }
+    protected void store(List<GenericMetrics> genericMetrics) {
+        Timer.Context c = metrics.getKairosDBTimer().time();
+        try {
+            kairosStore.store(genericMetrics);
+        } catch (Exception e) {
+            metrics.markKairosError();
+        } finally {
+            c.stop();
         }
     }
+
+}
 
 }
